% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utility_helpers.R
\name{add_delimiters}
\alias{add_delimiters}
\title{Add delimiters to text string}
\usage{
add_delimiters(string, delimiter)
}
\arguments{
\item{string}{the string to check for delimiter}

\item{delimiter}{a 1 or 2 character string containing the delimiter}
}
\value{
the string with the specified delimiters added (if necessary).
}
\description{
Add delimiters to text string
}
