% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{append_database_table}
\alias{append_database_table}
\title{Given a table from a database connection, append it to an existing table
using the INSERT INTO ... (COLUMN NAMES) SELECT ... FROM ... pattern.}
\usage{
append_database_table(
  db_connection,
  db,
  schema,
  tbl_name,
  list_of_columns,
  table_to_append
)
}
\description{
Like write_to_database, the original table must come from an SQL connection.
Does not append R tables onto SQL tables. You first need to write the R table
into SQL.
}
\details{
A common error occurs when character strings in the appended table exceed the
length of the existing varchar(n) limit in the original dataset.
E.g. appending the string 'ABCDE' into a varchar(3) column will error.
}
