% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_connect.R
\name{close_database_connection}
\alias{close_database_connection}
\title{Close an open database connection}
\usage{
close_database_connection(db_connection)
}
\arguments{
\item{db_connection}{the open connection to close.
Often created using \code{create_database_connection}.}
}
\description{
Good practice is to close open connections once they are finished with.
Large numbers of open connections can degrade performance.
}
