% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{collapse_indicator_columns}
\alias{collapse_indicator_columns}
\title{Many tables have mutually exclusive indicator columns.
This function provides a compact way to combine these columns to a single column
where the original columns all share the same prefix.}
\usage{
collapse_indicator_columns(input_tbl, prefix, yes_values, label = NA)
}
\description{
For example: sex_male = c("y", "n", "n"), sex_female = c("n", "y", "y")
Becomes: sex = c("male", "female", "female")
}
\details{
Where columns are not mutually exclusive, this function should not be used as
only a single value will be preserved in the output column and any duplication
will be removed invisably.

For example: eye_blue = c(1,1,0), eye_brown = c(1,1,1), eye_green = c(0,1,1)
Becomes eye = c("blue","blue","brown")
}
