% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{compress_table}
\alias{compress_table}
\title{Large SQL tables can be compressed to reduce the space they take up.}
\usage{
compress_table(db_connection, db, schema, tbl_name)
}
\description{
Large SQL tables can be compressed to reduce the space they take up.
}
