% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{copy_r_to_sql}
\alias{copy_r_to_sql}
\title{The inbuilt dbplyr copy_to function does not appear to work in our set-up.
Worse, it caused errors/locking, preventing other users who are using the
same connection method until their R session is restarted.}
\usage{
copy_r_to_sql(
  db_connection,
  db,
  schema,
  sql_table_name,
  r_table_name,
  OVERWRITE = FALSE
)
}
\description{
Hence we implement a established work around using the DBI package.
This is a revision of the original function (in previous versions),
and uses different functionality to simplify the process.
}
