% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{create_access_point}
\alias{create_access_point}
\title{Creates access point for R to run queries on SQL server.}
\usage{
create_access_point(db_connection, db, schema, tbl_name)
}
\description{
Use this in place of loading a table into R. It will be treated as an R dataframe
but the data will remain on the SQL server (instead of in R memory).
}
\details{
The same db_connection must be used foreach table access point. Otherwise
you will not be able to join tables together. Tables from different databases
within the same server can be accessed bythe same connection.

If table does not exist you will receive an error.
If you do not have permission to access to all the table columns you will receive an error.
Recommended solution: Create a View that only selects columns you have permission for.
}
