% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{create_database_connection}
\alias{create_database_connection}
\title{Any arguments passed to the function need to be named.
Default values are set at the top of this script.}
\usage{
create_database_connection(..., server = NA, database = NA, port = NA)
}
\description{
A single connection can access multiple associated databases.
We recommend using a single connection to access all tables
because attempts to join tables from different connections
(even if they are in the same database) perform poorly if
at all.
}
