% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{create_nonclustered_index}
\alias{create_nonclustered_index}
\title{Create a nonclustered index to improve table performance.
Unlike clustered indexes, multiple nonclustered indexes can be created.}
\usage{
create_nonclustered_index(db_connection, db, schema, tbl_name, index_columns)
}
\description{
Permanent high reuse tables may benefit from clustered indexes. But most researcher-created
tables are used a limited number of times and hence do not justify the additional overhead
of creating a clustered index. Non-clustered indexes are recommended in these cases.
}
