% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_write.R
\name{create_table}
\alias{create_table}
\title{Create new database table}
\usage{
create_table(
  db_connection,
  db = "[]",
  schema = "[]",
  tbl_name,
  named_list_of_columns,
  OVERWRITE = FALSE,
  query_path = NA
)
}
\arguments{
\item{db_connection}{the open connection to close.
Often created using \code{create_database_connection}.}

\item{db}{the name of the database. Can be different from the database used
when creating \code{db_connection} but must be on the same server. Warns if not
delimited in square brackets.}

\item{schema}{the name of the schema containing the Table or View. Warns if
not delimited in square brackets.}

\item{tbl_name}{the name of the Table in SQL. Warns is not delimited
in square brackets.}

\item{named_list_of_columns}{a named array, where the names are the column
names and the contents are the column data types. See details below.}

\item{OVERWRITE}{T/F should any existing table of the same name be deleted
first? Defaults to FALSE, and will error if a table with the same name
already exists.}

\item{query_path}{If provided will attempt to save a copy of the SQL code
sent to/executed on the database to the provided folder. Save occurs before
execution, hence useful for debugging.}
}
\value{
the result from executing the command using \code{DBI::dbExecute}
(seldom used).
}
\description{
Creates a new empty table in the database, overwriting an existing table if
instructed. Requires create table permissions.
}
\details{
Often used prior to \code{append_database_table}. Optional to use it prior to
\code{write_to_database}.

\code{named_list_of_columns} takes the format: \code{name = "sql type"}
For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{c(
  number_column_name = "[int] NOT NULL",
  date_column_name = "[date] NOT NULL",
  character_column_name = "[varchar](25) NULL"
)
}\if{html}{\out{</div>}}
}
