% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{create_view}
\alias{create_view}
\title{Returning connection to the new view}
\usage{
create_view(tbl_name, db_connection, db, schema, view_name, OVERWRITE = FALSE)
}
\description{
The view equivalent of write_to_database. Given a table from a database
connection, defines a new view with this definition.
}
\details{
The original table must come from an SQL connection.
Can only create views if the input connection is to the database containing
the views.
}
