% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_support.R
\name{db_schema}
\alias{db_schema}
\title{Append database and schema name}
\usage{
db_schema(db, schema)
}
\arguments{
\item{db}{the name of the database. Can be different from the database used
when creating \code{db_connection} but must be on the same server. Warns if not
delimited in square brackets.}

\item{schema}{the name of the schema containing the Table or View. Warns if
not delimited in square brackets.}
}
\value{
appended database and schema name
}
\description{
Some database connections can be used to access more than one database on the
same server. This requires that we explicitly include the database and schema
names in the connection.
}
\details{
This function joins the two together so they can be passed to
\code{dbplyr::in_schema}.
}
