% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{delete_table}
\alias{delete_table}
\title{Checks for existence of table/view prior to dropping.
Can only drop views if the input connection is to the database containing
the views.}
\usage{
delete_table(db_connection, db, schema, tbl_name, mode = "table")
}
\description{
Checks for existence of table/view prior to dropping.
Can only drop views if the input connection is to the database containing
the views.
}
