% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_connect.R
\name{display_connection_guidance}
\alias{display_connection_guidance}
\title{Print user guidance on database connections}
\usage{
display_connection_guidance()
}
\description{
Connecting to a database can be a frustrating task the first time around.
This is often a bespoke task that depends on the exact configuration of the
database.
}
\details{
This function returns example code that we found useful when connecting to,
and disconnecting from, a MS SQL Server. It is intended to serve as a guide
for arranging this in your own work.
}
