% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utility_helpers.R
\name{has_internal_delimiters}
\alias{has_internal_delimiters}
\title{Checks delimited string contains internal delimiters}
\usage{
has_internal_delimiters(string, delimiter)
}
\arguments{
\item{string}{the string to check for delimiter}

\item{delimiter}{a 1 or 2 character string containing the delimiter}
}
\value{
the string with the specified delimiters added (if necessary).
}
\description{
For most applications we want to error if an internal delimiter is found.
This helps protect against SL injection.
}
