% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utility_helpers.R
\name{no_special_characters}
\alias{no_special_characters}
\title{No special characters}
\usage{
no_special_characters(in_string)
}
\arguments{
\item{in_string}{a string to be checked for special characters.

No return. Silent pass if input is string and contains no special characters.
Otherwise errors if (1) input is not a string, (2) input has length greater
than 1, or (3) input string contains special characters.

Special characters are: \verb{;:'()\{\}?"*} and space.}
}
\description{
Check input string for absence of special characters
Helps prevent accidental SQL injection
}
