% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{pivot_table}
\alias{pivot_table}
\title{Takes label-value pair of columns and transforms table
so labels that were column entries become column names.}
\usage{
pivot_table(input_tbl, label_column, value_column, aggregator = "SUM")
}
\description{
SQL equivalent of tidyr::spread
}
\details{
Note that this tends to be an expensive operation.
So best not to include it as part of a long or complex
chain of commands.
}
