% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utility_helpers.R
\name{remove_delimiters}
\alias{remove_delimiters}
\title{Remove delimiters from text string}
\usage{
remove_delimiters(string, delimiter)
}
\arguments{
\item{string}{the string to check for delimiter}

\item{delimiter}{a 1 or 2 character string containing the delimiter}
}
\value{
the string with the specified delimiters removed (if necessary).
}
\description{
Remove delimiters from text string
}
