% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utility_helpers.R
\name{run_time_inform_user}
\alias{run_time_inform_user}
\title{Print to console time of function call followed by msg.}
\usage{
run_time_inform_user(msg, context = NA, print_off = FALSE)
}
\arguments{
\item{msg}{a message to display to the user with time stamps.}

\item{context}{optional, if provided formats the message. Three options:
{"heading", "details", "all"}}

\item{print_off}{if true nothing is printed. Allows for printing of an entire
super-function to be turned off.}
}
\description{
Print to console time of function call followed by msg.
}
