% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_support.R
\name{save_to_sql}
\alias{save_to_sql}
\title{Save SQL queries to files}
\usage{
save_to_sql(query, desc, path = getwd())
}
\arguments{
\item{query}{the text of the query to save. This may be generated using
\code{dbplyr::show_query} where required.}

\item{desc}{a description of the query. Use to name the file.}

\item{path}{the path to save the query. Will always be in a sub-folder named
"SQL tmp scripts" on this path. Creates the directory if required. Defaults
to the current working directory.}
}
\value{
The location and name of the saved file. If permissions do not exist
to save the file, the name and location of the attempted file.
}
\description{
For transparency and ease of debugging, we make a practice of saving to a
temporary folder all the SQL queries that write or change data on the server
(but not those that only fetch data).
}
\details{
This function provides a standardized way to save SQL code. As these scripts
are primarily intended to support debugging, they can be deleted without
concern.
}
