% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{set_connection_string}
\alias{set_connection_string}
\title{Description: dbplyr support functions for R development with SQL server}
\usage{
set_connection_string(server, database, port = NA)
}
\description{
Input: Connection details that must be manually set.
}
\details{
Output: support functions for ease of using R to manipulate SQL tables.

Author: Simon Anastasiadis

Dependencies: odbc, DBI, dplyr, dbplyr packages, utility_functions.R

Notes:
\itemize{
\item Connection details (line ~35) NOT FOR RELEASE!!
\item Some functions (and tests) assume SQL Server environment
due to differences in syntax between different implementations of SQL.
\item Uses code folding by headers (Alt + O to collapse all)
\item Example use included at end of file.
}

Issues:

History (reverse order):
2020-11-18 SA v2 for release
2020-07-12 SA addition of collapse indicator column function
2020-02-13 SA addition of pivot function
2020-01-09 SA v1 split dbplyr support out from utility_functions.R
#############################################################################

ODBC connections require a connection string to define the connection
between the database and the access point. This string has a default
format that this function satisfies.

This function does not need to be called directly to use the utility
functions. It is called by create_database_connection.
}
