% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{table_contains_required_columns}
\alias{table_contains_required_columns}
\title{Returns TRUE if the table contains all the required columns.
If only = TRUE, returns TRUE if the table ONLY contains all
the required columns.}
\usage{
table_contains_required_columns(tbl_to_check, required_columns, only = FALSE)
}
\description{
Checks column names only, does not consider contents.
}
