% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{table_or_view_exists_in_db}
\alias{table_or_view_exists_in_db}
\title{Returns true is table or view exists in the database
Used in place of 'dbExistsTable' and 'tbl \%in\% DBI::dbListTables(db_con)'
Because these approaches do not always handle schema consistently.}
\usage{
table_or_view_exists_in_db(db_connection, db, schema, tbl_name)
}
\description{
Returns true is table or view exists in the database
Used in place of 'dbExistsTable' and 'tbl \%in\% DBI::dbListTables(db_con)'
Because these approaches do not always handle schema consistently.
}
