% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{union_all}
\alias{union_all}
\title{Provides UNION ALL functionality from SQL. Appends two tables.}
\usage{
union_all(table_a, table_b, list_of_columns)
}
\description{
Requires a input tables to have identical column names. Provides as output
a single table the "union all" of all the input tables.
}
