% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{write_for_reuse}
\alias{write_for_reuse}
\title{Takes an R object (tbl_to_save) that is an SQL table that has been
manipulated/transformed and saves it as a new table in SQL.
Returns a connection to the new table.}
\usage{
write_for_reuse(
  db_connection,
  db,
  schema,
  tbl_name,
  tbl_to_save,
  index_columns = NA
)
}
\description{
For complex or extended analyses, this is recommended as it reduces
the complexity of the underlying SQL query that defines the manipulated
table.
}
