% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{write_to_database}
\alias{write_to_database}
\title{Returning connection to the new table}
\usage{
write_to_database(
  input_tbl,
  db_connection,
  db,
  schema,
  tbl_name,
  OVERWRITE = FALSE
)
}
\description{
Given a table from a database connection, writes to a new table using the
SELECT ... INTO ... FROM ... pattern.
}
\details{
Original table must come from an SQL connection.
Does not write R tables into SQL. Use copy_r_to_sql for this.

E.g. the following works
my_table <- create_access_point(....)
write_to_database(my_table, ...)
But this will not work
my_table <- data.frame(x = 1:10, y = 1:10)
write_to_database(my_table, ...)
}
