% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{close_database_connection}
\alias{close_database_connection}
\title{Good practice is to close open connections once they are finished with.
Large numbers of open connections can degrade performance.}
\usage{
close_database_connection(db_connection)
}
\description{
Good practice is to close open connections once they are finished with.
Large numbers of open connections can degrade performance.
}
