% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_manipulate.R
\name{collapse_indicator_columns}
\alias{collapse_indicator_columns}
\title{Combine indicator columns to single column}
\usage{
collapse_indicator_columns(input_tbl, prefix, yes_values, label = prefix)
}
\arguments{
\item{input_tbl}{a table or data frame. Works with local and remote tables.}

\item{prefix}{the prefix shared between all columns to combine}

\item{yes_values}{values that indicate the variable is true/active. Case-
sensitive. The result from combining different data types
(for example \code{c(1, "y")})  will depend on implicit conversions.}

\item{label}{an (optional) title for the new column. Defaults to the prefix.}
}
\value{
a remote (database) table where the multiple columns with the given
prefix have been replaced by a single categorical column.

Note that where columns are not mutually exclusive, this function should not
will remote duplication invisibly, only preserving a single value.

For example: suppose people can have multiple eye colors:
\verb{eye_blue = c(1,1,0), eye_brown = c(1,1,1), eye_green = c(0,1,1)}
Give output: \code{eye = c("blue","blue","brown")}
}
\description{
Many tables have mutually exclusive indicator columns. This function provides
a compact way to combine these columns to a single column where the original
columns all share the same prefix.
}
\details{
For example: \verb{sex_male = c("y", "n", "n"), sex_female = c("n", "y", "y")}
Becomes: \code{sex = c("male", "female", "female")}
}
