% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_manipulate.R
\name{compress_table}
\alias{compress_table}
\title{Compress a table}
\usage{
compress_table(db_connection, db, schema, tbl_name, query_path = NA)
}
\arguments{
\item{db_connection}{the open connection to close.
Often created using \code{create_database_connection}.}

\item{db}{the name of the database. Can be different from the database used
when creating \code{db_connection} but must be on the same server. Warns if not
delimited in square brackets.}

\item{schema}{the name of the schema containing the Table or View. Warns if
not delimited in square brackets.}

\item{tbl_name}{the name of the Table or View in SQL. Warns is not delimited
in square brackets.}

\item{query_path}{If provided will attempt to save a copy of the SQL code
sent to/executed on the database to the provided folder. Save occurs before
execution, hence useful for debugging.}
}
\value{
the result from executing the command using \code{DBI::dbExecute}
(seldom used).
}
\description{
Compacts an SQL Server table. Errors for non-SQL Server tables.
}
\details{
Large SQL tables can be compressed to reduce the space they take up.
We have observe this reduce storage space of tables by 50-75\% in many cases.
}
