% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{create_clustered_index}
\alias{create_clustered_index}
\title{Note: at most a single clustered index can be added to each table.
This operation is potentially expensive, so should be used only where needed.}
\usage{
create_clustered_index(db_connection, db, schema, tbl_name, index_columns)
}
\description{
For researcher created tables non-clustered indexes are recommended.
This function only provided for backwards compatibility.
}
