% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_manipulate.R
\name{create_nonclustered_index}
\alias{create_nonclustered_index}
\title{Add non-clustered index}
\usage{
create_nonclustered_index(db_connection, db, schema, tbl_name, index_columns)
}
\arguments{
\item{db_connection}{the open connection to close.
Often created using \code{create_database_connection}.}

\item{db}{the name of the database. Can be different from the database used
when creating \code{db_connection} but must be on the same server. Warns if not
delimited in square brackets.}

\item{schema}{the name of the schema containing the Table or View. Warns if
not delimited in square brackets.}

\item{tbl_name}{the name of the Table or View in SQL. Warns is not delimited
in square brackets.}

\item{index_columns}{an array containing the names of columns for the
index. Will error if columns are not found in table.}
}
\value{
the result from executing the command using \code{DBI::dbExecute}
(seldom used).
}
\description{
Adds non-clustered indexes to an SQL Server table. Errors for non-SQL Server
tables.
}
\details{
Create a non-clustered index to improve performance of joins.
Unlike clustered indexes, multiple non-clustered indexes can be created.

Clustered indexes tend to perform faster than non-clustered indexes.
Permanent high reuse tables may benefit from clustered indexes. But as most
researcher-created tables are used a limited number of times, they do not
justify the additional overhead of creating a clustered index. Hence
non-clustered indexes are recommended.
}
