% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{create_table}
\alias{create_table}
\title{Creates an empty table in the database, overwriting an existing copy if
instructed. Main use is to setup an empyt table that results can be iteratively
appended to.}
\usage{
create_table(
  db_connection,
  db,
  schema,
  tbl_name,
  named_list_of_columns,
  OVERWRITE = FALSE
)
}
\description{
Not necessary if creating a table directly from an existing table.
Use write_to_database for this.
}
\details{
named_list_of_columns takes the format name = "sql type"
For example:
list_of_columns <- (number = "\link{int} NOT NULL",
date = "\link{date} NOT NULL",
character = "\href{25}{varchar} NULL")
}
