% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_write.R
\name{create_view}
\alias{create_view}
\title{Create a View from an existing remote table}
\usage{
create_view(
  tbl_name,
  db_connection,
  db = "[]",
  schema = "[]",
  view_name,
  OVERWRITE = FALSE,
  query_path = NA
)
}
\arguments{
\item{tbl_name}{an existing remote table that defines/describes the new view}

\item{db_connection}{the open connection to close.
Often created using \code{create_database_connection}.}

\item{db}{the name of the database. Can be different from the database used
when creating \code{db_connection} but must be on the same server. Warns if not
delimited in square brackets.}

\item{schema}{the name of the schema containing the Table or View. Warns if
not delimited in square brackets.}

\item{view_name}{the name of the Table or View in SQL. Warns is not delimited
in square brackets.}

\item{OVERWRITE}{T/F should any existing view of the same name be deleted
first? Defaults to FALSE, and will error if a view with the same name already
exists.}

\item{query_path}{If provided will attempt to save a copy of the SQL code
sent to/executed on the database to the provided folder. Save occurs before
execution, hence useful for debugging.}
}
\value{
a connection to the new view (using \code{create_access_point})
}
\description{
The view equivalent of \code{write_to_database}. Given a table from a database
connection, defines a new View with this definition.
}
\details{
The original table must come from an SQL connection. Can only create views on
the database specified by the connection. Requires permission to create views
on this database.
}
