% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_manipulate.R
\name{pivot_table}
\alias{pivot_table}
\title{Pivot wider}
\usage{
pivot_table(input_tbl, label_column, value_column, aggregator = "SUM")
}
\arguments{
\item{input_tbl}{a remote (database) table}

\item{label_column}{the name of the column containing the labels}

\item{value_column}{the name of the column containing the values}

\item{aggregator}{where the labels are not unique, how should the multiple
values be resolved. Defaults to "SUM", but also works with any SQL
aggregation function such as "MAX", "MIN".}
}
\value{
a remote (database) table that has been pivoted wider.
}
\description{
Pivots long-thin tables to rectangular, taking a label-value pair of columns
and transforming table so labels that were cell values become column names.
}
\details{
SQL equivalent of tidyr::spread

Note that this tends to be an expensive operation. So best not to include as
part of a long or complex chains of commands.
}
