% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_helper_functions.R
\name{save_to_sql}
\alias{save_to_sql}
\title{All the SQL queries that write or change data on the server (but not those that
only fetch data) in these utility functions save an SQL code file of the command
that was executed.}
\usage{
save_to_sql(query, desc)
}
\description{
These scripts are primarily intended to support debugging.
They can be deleted without concern.
}
