% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_connect.R
\name{set_connection_string}
\alias{set_connection_string}
\title{ODBC connections require a connection string to define the connection
between the database and the access point. This string has a default
format that this function satisfies.}
\usage{
set_connection_string(server, database, port = NA)
}
\arguments{
\item{server}{}

\item{database}{}

\item{port}{}
}
\value{
the connection string to initiate a new database connection.
}
\description{
This function does not need to be called directly to use the utility
functions. It is called by create_database_connection.
}
\details{
Title
}
