% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_support.R
\name{table_contains_required_columns}
\alias{table_contains_required_columns}
\title{Confirm table contains the required columns}
\usage{
table_contains_required_columns(tbl_to_check, required_columns, only = FALSE)
}
\arguments{
\item{tbl_to_check}{}

\item{required_columns}{an array of character containing the names of the
required columns.}

\item{only}{if \code{only = TRUE}, checks that the table ONLY contains all the
required columns. Additional columns will cause the function to return FALSE.
Defaults to FALSE.}
}
\value{
T/F whether the (remote) table contains all the required columns.
}
\description{
Returns TRUE if the table contains all the required columns.
If only = TRUE, returns TRUE if the table ONLY contains all
the required columns.
}
\details{
Checks column names only, does not consider contents.
}
