% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_support.R
\name{table_or_view_exists_in_db}
\alias{table_or_view_exists_in_db}
\title{Check Table or View exists in the database}
\usage{
table_or_view_exists_in_db(db_connection, db = "[]", schema = "[]", tbl_name)
}
\arguments{
\item{db_connection}{the open connection to close.
Often created using \code{create_database_connection}.}

\item{db}{the name of the database. Can be different from the database used
when creating \code{db_connection} but must be on the same server. Warns if not
delimited in square brackets.}

\item{schema}{the name of the schema containing the Table or View. Warns if
not delimited in square brackets.}

\item{tbl_name}{the name of the Table or View in SQL. Warns is not delimited
in square brackets.}
}
\value{
T/F whether the Table or View exists
}
\description{
Returns TRUE if table or view exists in the database. Used in place of
\code{DBI::dbExistsTable} and \code{tbl \%in\% DBI::dbListTables(db_con)}
because these approaches do not always handle schema consistently.
}
