% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_manipulate.R
\name{union_all}
\alias{union_all}
\title{SQL Server Union all}
\usage{
union_all(table_a, table_b, list_of_columns)
}
\arguments{
\item{table_a}{a remote (database) table}

\item{table_b}{another remote (database) table}

\item{list_of_columns}{a list of columns found in both tables to be included
in the output table.}
}
\value{
a remote table containing the columns listed where the rows are
the combination of both tables.
}
\description{
Appends two tables, requires that both tables have identical column names.
}
