% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_support.R
\name{warn_if_missing_delimiters}
\alias{warn_if_missing_delimiters}
\title{Check for delimiters and special characters}
\usage{
warn_if_missing_delimiters(db, schema, tbl_name)
}
\arguments{
\item{db}{the name of the database. Can be different from the database used
when creating \code{db_connection} but must be on the same server.}

\item{schema}{the name of the schema containing the Table or View.}

\item{tbl_name}{the name of the Table or View in SQL.}
}
\description{
For clarity it is recommended that db, schema, and table names
be delimited using square brackets: "[]".
}
\details{
This function provides a standardized warning for missing delimiters to use
internally. It also checks for special characters in all three.
}
