% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dbplyr_write.R
\name{write_to_database}
\alias{write_to_database}
\title{Write database table back to database}
\usage{
write_to_database(
  input_tbl,
  db_connection,
  db = "[]",
  schema = "[]",
  tbl_name,
  OVERWRITE = FALSE,
  query_path = NA
)
}
\arguments{
\item{input_tbl}{an existing remote table that defines/describes the new
table to save}

\item{db_connection}{the open connection to close.
Often created using \code{create_database_connection}.}

\item{db}{the name of the database. Can be different from the database used
when creating \code{db_connection} but must be on the same server. Warns if not
delimited in square brackets.}

\item{schema}{the name of the schema containing the Table or View. Warns if
not delimited in square brackets.}

\item{tbl_name}{the name of the Table in SQL. Warns is not delimited
in square brackets.}

\item{OVERWRITE}{T/F should any existing table of the same name be deleted
first? Defaults to FALSE, and will error if a table with the same name
already exists.}

\item{query_path}{If provided will attempt to save a copy of the SQL code
sent to/executed on the database to the provided folder. Save occurs before
execution, hence useful for debugging.}
}
\value{
a connection to the new table (using \code{create_access_point})
}
\description{
In R it is straightforward, to load a csv file, undertake some manipulation
and then save out a new csv file. However, when working with databases, this
is not as common a pattern for analysis.
}
\details{
Given a table from a database connection, writes to a new table using the
SELECT ... INTO ... FROM ... pattern.

Intended to be used in a Sandpit environment. Will error is user does not
have create table or insert into permissions.

Does not write R tables into SQL. The \code{input_tbl} must be a remote data
frame. Use \code{copy_r_to_sql} if the input table is a local R data frame.
}
